Rename dpt_test files to closer match dpt package
