+++ test/dpt_tests/dpt_16_string_test.py
