+++ test/dpt_tests/dpt_6_test.py
